Get regression api working
